Refactor file paths to use proper casing
