+++ src/Structures/Utils.ts
