+++ src/Utils/NodeCheck.ts
@@ -1,4 +1,4 @@
-import { NodeOptions } from "../structures/Node";
+import { NodeOptions } from "../Structures/Node";
 
 export default function NodeCheck(options: NodeOptions) {
 	if (!options) throw new TypeError("NodeOptions must not be empty.");
