+++ src/index.ts
@@ -1,11 +1,11 @@
-export * from "./structures/Manager";
-export * from "./structures/Node";
-export * from "./structures/Player";
-export * from "./structures/Queue";
-export * from "./structures/Utils";
-export * from "./structures/Filters";
-export * from "./structures/Rest";
-export * from "./utils/FiltersEqualizers";
-export * from "./utils/ManagerCheck";
-export * from "./utils/NodeCheck";
-export * from "./utils/PlayerCheck";
+export * from "./Structures/Manager";
+export * from "./Structures/Node";
+export * from "./Structures/Player";
+export * from "./Structures/Queue";
+export * from "./Structures/Utils";
+export * from "./Structures/Filters";
+export * from "./Structures/Rest";
+export * from "./Utils/FiltersEqualizers";
+export * from "./Utils/ManagerCheck";
+export * from "./Utils/NodeCheck";
+export * from "./Utils/PlayerCheck";
